Move `StateValidatorData` and statics to json types package
